+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add onboarding flow
 - Copied MessageCard and MessageButton SwiftUI view from Planetary renaming Message to Note
 - Added a basic profile screen
+- Added AvatarView for rendering profile pictures on NoteCards
 
 ## [0.1 Build 1]